Update python-mpmath to 1.1.0 and build it with python3
Summary:
Full changelog [here](https://github.com/fredrik-johansson/mpmath/blob/master/CHANGES)

**Package changes:**
`python-mpmath` and `python3-mpmath` have the same source and they are both dependencies for `sympy`.
Python2 support for `sympy` is being removed, so `python-mpmath` which is built with python2 have to be deprecated.
But to have python names consistency `python-mpmath` is being built with python3 and `python3-mpmath` is going to be deprecated.

Depends on D9410

Signed-off-by: Algent Albrahimi <algent@protonmail.com>

Test Plan:
- Update and build `sympy`.
- Run the examples:
```
>>> from mpmath import *
>>> mp.dps = 50
>>> print(mpf(2) ** mpf('0.5'))
1.4142135623730950488016887242096980785696718753769
>>> print(2*pi)
6.2831853071795864769252867665590057683943387987502
>>> mp.dps = 100
>>> print(mpf(2) ** mpf('0.5'))
1.414213562373095048801688724209698078569671875376948073176679737990732478462107038850387534327641573
>>> print(2*pi)
6.283185307179586476925286766559005768394338798750211641949889184615632812572417997256069650684234136
>>>
```

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D9411